// packages/cosmwasm-utils/src/client/types/Multiquery.types.ts
/**
 * This file was automatically generated by @abstract-money/ts-codegen@0.37.0-beta-3.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run the @abstract-money/ts-codegen generate command to regenerate this file.
 */

// biome-ignore lint/suspicious/noEmptyInterface: <explanation>
export interface InstantiateMsg {}
export type ExecuteMsg = string
export type QueryMsg =
  | {
      contract_version: {}
    }
  | {
      aggregate: {
        queries: Call[]
      }
    }
  | {
      try_aggregate: {
        include_cause?: boolean | null
        queries: Call[]
        require_success?: boolean | null
      }
    }
  | {
      try_aggregate_optional: {
        include_cause?: boolean | null
        queries: CallOptional[]
      }
    }
  | {
      block_aggregate: {
        queries: Call[]
      }
    }
  | {
      block_try_aggregate: {
        include_cause?: boolean | null
        queries: Call[]
        require_success?: boolean | null
      }
    }
  | {
      block_try_aggregate_optional: {
        include_cause?: boolean | null
        queries: CallOptional[]
      }
    }
export type Addr = string
export type Binary = string
export interface Call {
  address: Addr
  data: Binary
}
export interface CallOptional {
  address: Addr
  data: Binary
  require_success: boolean
}
export interface AggregateResult {
  return_data: CallResult[]
}
export interface CallResult {
  data: Binary
  success: boolean
}
export interface BlockAggregateResult {
  block: number
  return_data: CallResult[]
}
export interface ContractVersion {
  contract: string
  version: string
}